Move reusable code into a janeswalk module
