+++ modules/janeswalk/janeswalk.js
@@ -1,0 +1,3 @@
+function Hello() { return "Hello World!"; }
+
+export default Hello;
